requirements.txt
playwright==1.48.0
pyyaml==6.0.2
aiofiles==24.1.0
